implemented worker pools instead of semaphore. also slight refactor.
