+++ column-prefetching-server/internal/service/prefetching_service.go
@@ -11,32 +11,6 @@ import (
 	"sync"
 )
 
-type PrefetchingService struct {
-	S3Service    *S3Service
-	CacheService *CacheService
-	Config       project_config.PrefetchingConfig
-}
-
-type PrefetchRequest struct {
-	Bucket  string
-	Prefix  string
-	Columns []string
-}
-
-type RequestedColumn struct {
-	ColumnName string
-	Start      int64
-	End        int64
-}
-type ParquetColumnData struct {
-	Bucket string
-	Key    string
-	Column string
-	Data   []byte
-	Etag   string
-	Range  string
-}
-
 func NewPrefetchingService(
 	s3Service *S3Service,
 	cacheService *CacheService,
@@ -67,11 +41,13 @@ func (service *PrefetchingService) PrefetchColumns(ctx context.Context, req Pref
 		columnSet[column] = struct{}{}
 	}
 
-	// TODO: throughout the code, the semaphore pattern is causing serious memory issues. Try switching to worker pool.
+	jobs := make(chan FileJob, len(files))
 
-	// creating a buffered channel, of length concurrencyLimit to act as a semaphore. This limits the number of concurrent goroutines.
-	sem := make(chan struct{}, service.Config.ConcurrencyLimit)
 	var wg sync.WaitGroup
+	for i := 0; i < service.Config.ConcurrencyLimit; i++ {
+		wg.Add(1)
+		go service.fileWorker(ctx, jobs, &wg)
+	}
 
 	for _, file := range files {
 
@@ -79,24 +55,13 @@ func (service *PrefetchingService) PrefetchColumns(ctx context.Context, req Pref
 			continue
 		}
 
-		wg.Add(1)
-
-		go func(file types.Object) {
-			defer wg.Done()
-
-			// acquire a resource by placing a value in the channel
-			sem <- struct{}{}
-
-			// release the resource when done
-			defer func() { <-sem }()
-
-			err := service.prefetchFileColumns(ctx, req.Bucket, file, columnSet)
-			if err != nil {
-				log.Printf("failed to process parquet file %q: %v", file, err)
-			}
-		}(file)
+		jobs <- FileJob{
+			Bucket:    req.Bucket,
+			File:      file,
+			ColumnSet: columnSet,
+		}
 	}
-
+	close(jobs)
 	wg.Wait()
 
 	fmt.Printf("Total sequential time spent making S3 Requests: %d seconds \n", GetTotalS3CPUTime())
@@ -111,31 +76,26 @@ func (service *PrefetchingService) prefetchFileColumns(ctx context.Context, buck
 
 	requestedColumns, _ := getRequestedColumns(footerData, columnSet)
 
-	sem := make(chan struct{}, service.Config.ConcurrencyLimit)
+	jobs := make(chan ColumnJob, len(requestedColumns))
+
 	var wg sync.WaitGroup
+	for i := 0; i < service.Config.ConcurrencyLimit; i++ {
+		wg.Add(1)
+		go service.columnWorker(ctx, jobs, &wg)
+	}
 
 	for _, requestedColumn := range requestedColumns {
-		wg.Add(1)
 
-		go service.prefetchColumn(ctx, bucket, *file.Key, requestedColumn, sem, &wg)
+		jobs <- ColumnJob{
+			Bucket:          bucket,
+			FileKey:         *file.Key,
+			RequestedColumn: requestedColumn,
+		}
 	}
 
 	wg.Wait()
 
 	return nil
-
-}
-
-// prefetchColumn is responsible for sending the required work to the S3Service and CacheService to prefetch the
-// requested column data from the parquet file, storing it in the cache.
-func (service *PrefetchingService) prefetchColumn(ctx context.Context, bucket string, fileKey string, requestedColumn RequestedColumn, sem chan struct{}, wg *sync.WaitGroup) {
-	defer wg.Done()
-
-	sem <- struct{}{}
-	defer func() { <-sem }()
-
-	columnData, _ := service.S3Service.GetColumnData(ctx, bucket, fileKey, requestedColumn)
-	service.CacheService.CacheColumnData(columnData)
 }
 
 // getRequestedColumns is responsible for extracting the required column data as determined by the initial HTTP request.
